chrome: merge eslint configs
