+++ packages/recorder/vite.config.ts
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
+import path from 'path';
+
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
-import path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
