+++ packages/trace-viewer/src/sw/traceModelBackends.ts
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import type * as zip from '@zip.js/zip.js';
 // @ts-ignore
 import * as zipImport from '@zip.js/zip.js/lib/zip-no-worker-inflate.js';
+
+import type * as zip from '@zip.js/zip.js';
 import type { TraceModelBackend } from './traceModel';
 
 const zipjs = zipImport as typeof zip;
