+++ packages/trace-viewer/vite.config.ts
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
+import path from 'path';
+
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+
 import { bundle } from './bundle';
-import path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
