+++ packages/web/src/theme.ts
@@ -15,6 +15,7 @@
  */
 
 import React from 'react';
+
 import { settings } from './uiUtils';
 
 declare global {
